requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
backports.tarfile==1.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
dataclasses-json==0.6.7
datasets==2.14.7
dill==0.3.7
distro==1.9.0
dspy==0.1.5
dspy-ai==2.4.5
exceptiongroup==1.2.2
fastapi==0.112.2
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
itsdangerous==2.2.0
jaraco.context==6.0.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-community==0.2.14
langchain-core==0.2.36
langchain-text-splitters==0.2.2
langsmith==0.1.106
latex2mathml==3.77.0
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.43.0
optuna==3.6.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
sympy==1.13.2
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.4
wolframalpha==5.1.3
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.9.4
